Allow finer pitch bends

+++ Src/Actions/AddControlEvent.cpp
@@ -27,7 +27,7 @@ using namespace AriaMaestosa::Action;
 
 // ---------------------------------------------------------------------------------------------------------
 
-AddControlEvent::AddControlEvent(const int x, const int value, const int controller) :
+AddControlEvent::AddControlEvent(const int x, const float value, const int controller) :
 //I18N: (undoable) action name
 SingleTrackAction( _("add control event") )
 {
